backend/requirements.txt
python-dotenv==1.2.1
Flask==3.1.2
pymongo==4.16.0
pandas==3.0.0
plotly==6.5.2
scikit-learn==1.8.0
Werkzeug==3.1.5
flask-cors==6.0.2
huggingface-hub==0.26.5
requests==2.32.3